Fix bug when receiving messages
This was a result of allocating the received message in the stack. The
stack scope was limited to the consumer and by the time it rendered,
that data had been changed.

By converting the Message model into a Q_OBJECT it is easier to allocate
on the heap and deallocate when possible with `msg->deleteLater()`.